add update logic

+++ controllers/character.js
@@ -36,12 +36,14 @@ const getCharacterById = async (req, res) => {
 const addCharacter = async (req, res) => {
   const images = [];
 
-  for (const file of req.files) {
-    const { url } = await cloudinary.uploader.upload(file.path, {
-      folder: "characters",
-    });
-    images.push(url);
-    await fs.unlink(file.path);
+  if (req.files && req.files.length > 0) {
+    for (const file of req.files) {
+      const { url } = await cloudinary.uploader.upload(file.path, {
+        folder: "characters",
+      });
+      images.push(url);
+      await fs.unlink(file.path);
+    }
   }
 
   const characterData = {
@@ -51,6 +53,44 @@ const addCharacter = async (req, res) => {
 
   const result = await Character.create(characterData);
   res.status(201).json(result);
+};
+
+const updateCharacter = async (req, res) => {
+  const { id } = req.params;
+
+  const existingCharacter = await Character.findById(id);
+
+  if (!existingCharacter) {
+    throw HttpError(404, `Character with id ${id} not found`);
+  }
+
+  const images = [];
+
+  if (req.files && req.files.length > 0) {
+    for (const file of req.files) {
+      const { url } = await cloudinary.uploader.upload(file.path, {
+        folder: "characters",
+      });
+      images.push(url);
+      await fs.unlink(file.path);
+    }
+  }
+
+  const characterData = {
+    ...existingCharacter.toObject(),
+    ...req.body,
+    ...(images.length > 0 && { images }),
+  };
+
+  const result = await Character.findByIdAndUpdate(id, characterData, {
+    new: true,
+  });
+
+  if (!result) {
+    throw HttpError(404, `Character with id ${id} not found`);
+  }
+
+  res.json(result);
 };
 
 const deleteCharacter = async (req, res) => {
@@ -68,6 +108,6 @@ module.exports = {
   getAllCharacters: ctrlWrapper(getAllCharacters),
   getCharacterById: ctrlWrapper(getCharacterById),
   addCharacter: ctrlWrapper(addCharacter),
-
+  updateCharacter: ctrlWrapper(updateCharacter),
   deleteCharacter: ctrlWrapper(deleteCharacter),
 };
